backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-socketio==5.3.5
yt-dlp==2024.3.10
python-socketio==5.11.0
apify-client==1.7.0
requests==2.31.0
google-generativeai==0.8.3
python-dotenv==1.0.0
